Better name for Search bar view used in SwiftUI
